ai_backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
sentence-transformers>=2.6.0
faiss-cpu>=1.10.0
numpy>=1.26.0
pydantic>=2.7.0
openai>=1.30.0
supabase>=2.5.0
python-dotenv>=1.0.1
aiohttp>=3.9.0
better-profanity>=0.7.0
nltk>=3.9.0
requests>=2.31.0
python-multipart>=0.0.9
transformers>=4.41.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Optional local TTS fallback (if you prefer Coqui over Piper CLI)
# TTS>=0.22.0
imageio-ffmpeg>=0.5.1
